Improved JpaMapping and Query Optimization

+++ src/main/java/com/namastetractors/namaste_tractors_backend/dto/tractordto/TractorResponseDto.java
@@ -2,6 +2,7 @@ package com.namastetractors.namaste_tractors_backend.dto.tractordto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -10,12 +11,11 @@ public class TractorResponseDto {
     private Long id;
     private String model;
     private int hp;
-    private double price;
+    private BigDecimal price;
 
     private String brand;
 
     private TractorSpecDto specification;
 
     private List<ImageResponseDto> images;
-
 }
